+++ changelog.md
@@ -1,1 +1,1 @@
-- Updated to 1.16.5
+- Added additional data to image